+++ packages/pn-personafisica-webapp/src/pages/Contacts.page.tsx
@@ -1,7 +1,8 @@
 import { useCallback, useEffect, useMemo, useState } from 'react';
 import { Trans, useTranslation } from 'react-i18next';
+import { useNavigate } from 'react-router-dom';
 
-import { Box, Link, Stack } from '@mui/material';
+import { Box, Button, Link, Stack } from '@mui/material';
 import { ApiErrorWrapper, TitleBox } from '@pagopa-pn/pn-commons';
 
 import ContactsSummaryCards from '../components/Contacts/ContactsSummaryCards';
@@ -11,6 +12,7 @@ import SpecialContacts from '../components/Contacts/SpecialContacts';
 import LoadingPageWrapper from '../components/LoadingPageWrapper/LoadingPageWrapper';
 import { PFEventsType } from '../models/PFEventsType';
 import { FAQ_WHAT_IS_AAR, FAQ_WHAT_IS_COURTESY_MESSAGE } from '../navigation/externalRoutes.const';
+import { DIGITAL_DOMICILE_ACTIVATION } from '../navigation/routes.const';
 import { CONTACT_ACTIONS, getDigitalAddresses } from '../redux/contact/actions';
 import { contactsSelectors, resetState } from '../redux/contact/reducers';
 import { useAppDispatch, useAppSelector } from '../redux/hooks';
@@ -23,6 +25,7 @@ const Contacts = () => {
   const addressesData = useAppSelector(contactsSelectors.selectAddresses);
   const [pageReady, setPageReady] = useState(false);
   const { LANDING_SITE_URL } = getConfiguration();
+  const navigate = useNavigate();
 
   const showSpecialContactsSection = addressesData.specialAddresses.length > 0;
 
@@ -92,6 +95,9 @@ const Contacts = () => {
           subTitle={subtitle}
           variantSubTitle={'body1'}
         />
+        <Button onClick={() => navigate(DIGITAL_DOMICILE_ACTIVATION)} variant="contained">
+          {t('button.refresh')}
+        </Button>
         <ApiErrorWrapper
           apiId={CONTACT_ACTIONS.GET_DIGITAL_ADDRESSES}
           reloadAction={fetchAddresses}
